+++ tracky-consumer/src/main/java/kernel360/trackyconsumer/application/service/ConsumerService.java
@@ -1,19 +1,20 @@
 package kernel360.trackyconsumer.application.service;
 
 import java.time.LocalDateTime;
-import java.time.ZoneId;
 import java.util.List;
 
-import jakarta.transaction.Transactional;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import kernel360.trackyconsumer.application.dto.CarOnOffRequest;
 import kernel360.trackyconsumer.application.dto.CycleGpsRequest;
+import kernel360.trackyconsumer.application.dto.GpsHistoryMessage;
 import kernel360.trackyconsumer.infrastructure.repository.CarEntityRepository;
 import kernel360.trackyconsumer.infrastructure.repository.DriveRepository;
 import kernel360.trackyconsumer.infrastructure.repository.GpsHistoryRepository;
 import kernel360.trackyconsumer.infrastructure.repository.LocationEntityRepository;
 import kernel360.trackyconsumer.infrastructure.repository.RentEntityRepository;
-import kernel360.trackyconsumer.config.RabbitMQConfig;
-import kernel360.trackyconsumer.application.dto.GpsHistoryMessage;
 import kernel360.trackycore.core.common.entity.CarEntity;
 import kernel360.trackycore.core.common.entity.DriveEntity;
 import kernel360.trackycore.core.common.entity.GpsHistoryEntity;
@@ -22,38 +23,43 @@ import kernel360.trackycore.core.common.entity.RentEntity;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-import org.springframework.amqp.rabbit.annotation.RabbitListener;
-import org.springframework.messaging.handler.annotation.Header;
-import org.springframework.messaging.handler.annotation.Payload;
-import org.springframework.stereotype.Service;
-
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class CarInfoConsumer {
+public class ConsumerService {
 
-	private final DriveRepository driveRepository;
 	private final GpsHistoryRepository gpsHistoryRepository;
+	private final DriveRepository driveRepository;
 	private final LocationEntityRepository locationEntityRepository;
 	private final RentEntityRepository rentEntityRepository;
 	private final CarEntityRepository carEntityRepository;
 
-	@RabbitListener(queues = RabbitMQConfig.ONOFF_QUEUE_NAME)
+	@Async
 	@Transactional
-	public void receiveCarOnOffMessage(@Payload CarOnOffRequest message,
-		@Header("amqp_receivedRoutingKey") String routingKey) {
+	public void receiveCycleInfo(GpsHistoryMessage request) {
 
-		log.info("메시지 수신: {}", message.toString());
-		switch (routingKey) {
-			case "on":
-				processOnMessage(message);
-				break;
-			case "off":
-				processOffMessage(message);
-				break;
+		// 처리 로직
+		List<CycleGpsRequest> cycleGpsRequestList = request.getCList();
+
+		DriveEntity drive = driveRepository.findByMdnAndOtime(request.getMdn(), request.getOTime());
+
+		// GPS 쪼개서 정보 저장
+		for (int i = 0; i < request.getCCnt(); i++) {
+			saveCycleInfo(drive, request.getOTime(), cycleGpsRequestList.get(i).getSum(), cycleGpsRequestList.get(i));
 		}
 	}
 
+	private void saveCycleInfo(DriveEntity drive, LocalDateTime oTime, double sum, CycleGpsRequest cycleGpsRequest) {
+		// public void saveGpsMessage(LocalDateTime oTime, double sum, CycleGpsRequest cycleGpsRequest) {
+
+		long maxSeq = gpsHistoryRepository.findMaxSeqByDrive(drive);
+		GpsHistoryEntity gpsHistoryEntity = cycleGpsRequest.toGpsHistoryEntity(maxSeq + 1, drive, oTime, sum);
+
+		gpsHistoryRepository.save(gpsHistoryEntity);
+
+	}
+
+	@Transactional
 	public void processOnMessage(CarOnOffRequest carOnOffRequest) {
 
 		LocationEntity location = carOnOffRequest.toLocationEntity();
@@ -65,7 +71,7 @@ public class CarInfoConsumer {
 		RentEntity rent = rentEntityRepository.findMyMdnAndTime(mdn, carOnOffRequest.getOnTime());
 
 		DriveEntity drive = DriveEntity.create(mdn, rent.getRentUuid(), car.getDevice().getId(), location,
-		carOnOffRequest.getOnTime());
+			carOnOffRequest.getOnTime());
 
 		// DriveEntity drive = DriveEntity.create("temp1234", "some_uuid", 1, location,
 		// 	carOnOffRequest.getOnTime());
@@ -73,6 +79,7 @@ public class CarInfoConsumer {
 		driveRepository.save(drive);
 	}
 
+	@Transactional
 	public void processOffMessage(CarOnOffRequest carOnOffRequest) {
 
 		/** 시동 off시 다음 동작 수행
@@ -92,33 +99,5 @@ public class CarInfoConsumer {
 
 		LocationEntity location = drive.getLocation();
 		location.updateEndLocation(carOnOffRequest.getLat(), carOnOffRequest.getLon());
-	}
-
-	// GPS 정보 처리 큐
-	@RabbitListener(queues = RabbitMQConfig.GPS_QUEUE_NAME)
-	@Transactional
-	public void receiveCarMessage(GpsHistoryMessage message) {
-
-		List<CycleGpsRequest> cycleGpsRequestList = message.getCList();
-
-		DriveEntity drive = driveRepository.findByMdnAndOtime(message.getMdn(), message.getOTime());
-
-		// GPS 쪼개서 정보 저장
-		for (int i = 0; i < message.getCCnt(); i++) {
-			saveGpsMessage(drive, message.getOTime(), cycleGpsRequestList.get(i).getSum(), cycleGpsRequestList.get(i));
-			// saveGpsMessage(message.getOTime(), cycleGpsRequestList.get(i).getSum(), cycleGpsRequestList.get(i));
-		}
-
-	}
-
-	public void saveGpsMessage(DriveEntity drive, LocalDateTime oTime, double sum, CycleGpsRequest cycleGpsRequest) {
-	// public void saveGpsMessage(LocalDateTime oTime, double sum, CycleGpsRequest cycleGpsRequest) {
-
-		long maxSeq = gpsHistoryRepository.findMaxSeqByDrive(drive);
-		GpsHistoryEntity gpsHistoryEntity = cycleGpsRequest.toGpsHistoryEntity(maxSeq + 1, drive, oTime, sum);
-		log.info("gpshitory 엔티티: {}", gpsHistoryEntity.toString());
-
-		gpsHistoryRepository.save(gpsHistoryEntity);
-
 	}
 }
